Stepper working with changing out components based on what step we are on.
