+++ eventkit_cloud/ui/static/ui/app/components/CreateExport.js
@@ -8,9 +8,8 @@ import MenuItem from 'material-ui/MenuItem'
 import BreadcrumbStepper from './BreadcrumbStepper'
 import MoreVertIcon from 'material-ui/svg-icons/navigation/more-vert'
 import primaryStyles from '../styles/constants.css'
-import ExportAOI, {MODE_DRAW_BBOX, MODE_NORMAL} from './ExportAOI'
-import {updateBbox} from '../actions'
-import {toggleDrawCancel, toggleDrawRedraw} from '../actions/drawToolBarActions.js'
+
+
 import FlatButton from 'material-ui/FlatButton';
 import RaisedButton from 'material-ui/RaisedButton';
 import {
@@ -25,7 +24,7 @@ class CreateExport extends React.Component {
 
     constructor() {
         super()
-        this._handleBoundingBoxChange = this._handleBoundingBoxChange.bind(this)
+        
     }
 
 
@@ -67,10 +66,7 @@ class CreateExport extends React.Component {
                                             </IconMenu>}
                 />
                 <BreadcrumbStepper/>
-                <div>
-                    <ExportAOI mode={this._mapMode} 
-                               onBoundingBoxChange={() => this._handleBoundingBoxChange()}/>
-                </div>
+                
                 <div >
                     {this.props.children}
                 </div>
@@ -81,22 +77,7 @@ class CreateExport extends React.Component {
         );
     }
 
-    //
-    // Internal API
-    //
-
-    get _mapMode() {
-        if (this.props.location === 'exportAOI') {
-            return MODE_DRAW_BBOX
-        }
-        return MODE_NORMAL
-    }
-
-    _handleBoundingBoxChange(bbox) {
-        console.log('Running Handle bounding box change in CreateExport.js')
-        console.log(this.props.bbox)
-    }
-
+    
 }
